:art: Move tab bar component
