+++ src/components/TabBar/index.tsx
@@ -2,32 +2,22 @@ import { BorderTab } from '@cos-ui/react';
 import { useLayoutEffect, useState } from 'react';
 import { Outlet, useLocation, useNavigate } from 'react-router-dom';
 
-const TAB_LIST = [
-  {
-    id: 1,
-    label: '소개',
-    link: 'introduce',
-  },
-  {
-    id: 2,
-    label: '공지사항',
-    link: 'announcement',
-  },
-  {
-    id: 3,
-    label: '아티클',
-    link: 'article',
-  },
-];
+type TabBarProps = {
+  tabs: {
+    id: number;
+    label: string;
+    link: string;
+  }[];
+};
 
-const TabBar = () => {
+const TabBar = ({ tabs }: TabBarProps) => {
   const navigate = useNavigate();
   const location = useLocation();
 
   const [defaultIndex, setDefaultIndex] = useState(0);
 
   useLayoutEffect(() => {
-    const index = TAB_LIST.findIndex(({ link }) =>
+    const index = tabs.findIndex(({ link }) =>
       location.pathname.includes(link),
     );
     setDefaultIndex(index);
@@ -37,7 +27,7 @@ const TabBar = () => {
     <div>
       <BorderTab.Group defaultIndex={defaultIndex}>
         <BorderTab.List>
-          {TAB_LIST.map(({ id, label, link }) => (
+          {tabs.map(({ id, label, link }) => (
             <BorderTab key={id} onClick={() => navigate(link)}>
               {label}
               <BorderTab.ActiveBar />
